requirements.txt
# For Daraz scraper (main.py, detail_main.py)
crawl4ai
python-dotenv

# For StarTech scraper (startech_scraper.py)
firecrawl-py
beautifulsoup4
tenacity
nest_asyncio
# crawl4ai is already listed

# For Corpus builders
# (No extra libraries needed beyond standard Python)

# For RAG App (app.py)
streamlit
openai
rank_bm25
chonkie